+++ packages/core/src/lib/static/plugins/index.ts
@@ -2,5 +2,5 @@
  * @file Automatically generated by barrelsby.
  */
 
-export * from './CopyPlugin';
+export * from './ViewPlugin';
 export * from './getStaticPlugins';
